Jamdo/RessourceGathering/requirements.txt
mwparserfromhell-0.5.3
sqlalchemy
flask